+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - [**BC**] Consume with error handling
 - Fix to work without a tracing
 - [**BC**] Remove `Logger` type and use `ILogger` from `Microsoft.Extensions.Logging`
+- [**BC**] Add configuration option to set `CommitMessage` *mode*
 
 ## 16.2.0 - 2021-09-14
 - Update dependencies